% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Plot random walk chain}
\usage{
main(p.length, p.prob, p.num, p.seed = 1)
}
\arguments{
\item{p.length}{length of the generated chain}

\item{p.prob}{probability to win in one game}

\item{p.num}{...}

\item{p.seed}{random seed}
}
\value{
No return but plot the figure.
}
\description{
Make it easy to plot. Thank you for using it.
}
\examples{
main(100,0.5,5,1)
main(100,0.51,5,1)
main(100,0.4,5,1)
main(100,0.6,5,1)
}
