% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomwalkli.R
\docType{package}
\name{randomwalklia}
\alias{randomwalklia}
\title{test :A package for test}
\description{
How to use the package:
}
\section{Function main}{

The key function is function main ..........
}

